gtfs renderer/requirements.txt
Flask>=2.0.1
requests>=2.22.0
Pillow>=8.3.2
